Stop using lazy_static for the histogram energies/boundaries
Instead use `static mut` and some unsafe code, and initialize once when
creating the history the first time.

This removes initialization checks on each access to the arrays and
overall speeds up the addition of items to the histogram by a factor
of 1.3.

Compared to the Rust version it is now only about 3% slower.