Change how the sample peak is calculated
This is a preparation for adding support for planar/non-interleaved
inputs. The outer loop now goes over all channels instead of going over
all frames.

This also speeds up sample peak measurement by about 3.8% and it's now
slightly faster than the C code. Before it was the only part of the code
that was measurably slower by about 3%.