Remove C comparison tests and benchmarks
The implementation is proven to be equivalent to the C implementation
now and faster in any meaningful way so simplify the build system and
get rid of the C code.

Next various unit tests have to be added for the removed comparison
tests to ensure no regressions in the future.

Fixes https://github.com/sdroege/ebur128/issues/37