history: Drop TODO about list-based queues
I got curious, and tested it. Did not try linked list of arrays, but a
a simpler Vec-of-Vec approach, and focused on speed of insertion and
iteration. (To gauge effects of allocation-size)

Did try for inner-vec-sizes of 64..131072. 2048 turned out to be a peak.
Increasing over this gave no significant benefit. Also did compared to
pure vector-approach, for durations of 1H - 10000H of history. At smaller
time-periods, the simple vector out-performed the 2-layered approach
significantly. (~2x) At larger time-periods, the simple-vector still won,
but with a lesser gap.

Also noted; with 100H hours measurement-window, the insert and iterate-time
on my machine is only 22ms, which should be a really tiny amount compared
to everything else done to analyse 100 hours of audio.

All-in-all, I doubt there is much to win by a list-of-arrayed approach, and
I think we could drop the TODO.